Java Project - Quiz Maker
This program will be used by the Quiz Master to create quizzes. The Quiz Master will be 
allowed to input a set of questions about a subject. For each question the Quiz Master will 
also input 4 answers, and mark which of them are correct (for each question, 1 or more 
answers are correct). The Quiz Master can then export the quiz in a text file. The quiz can be 
opened later by the Quiz Maker for editing