C# Project - Pizza Menu desktop Application
Firehouse Pizza is a new fast-food outlet employing a novel business 
concept - customisable pizzas! You are requested to develop software 
to support the operations of this new shop.
Menus will lead the customer to choose the pizza size, sauce, crust,
crust filling (if applicable), and toppings. 
- Pizza size could be small, medium, and large.
- Pizza sauce can be tomato or cream.
- Pizza crust can be regular or stuffed. In case of stuffed crust, 
the customer would need to choose a crust filling, which can 
be mozzarella or spiced cheese.
- Each pizza shall have five different toppings. Hot and spicy 
toppings have an associated grade which can be mild, moderate, and ferocious.
The price of each pizza is determined by summing up the prices for the pizza size and its toppings. An extra cost 
for the crust filling is added where applicable.
Once all the pizza details are chosen, a summary screen should show an itemised receipt with the total price as 
shown below. A warning of how hot and spicy the pizza is should also show up when necessary. This value is not 
cumulative, and is determined by the most hot and spicy topping chosen.