SQL project - Developing DBMS Applications
Scenario - A local video game company has produced a game that allows gamers to generate revenue whilst 
playing their game. They have decided that it was time to generate statistics about the gamers and 
their balances. They have generated a JSON file with the client’s details (persondata.json). 
The main aim is to load this JSON file into a database, making sure it meets the given design. Then 
automatically generate statistics for the uploaded data. A front-end will be required, this can be coded 
either in Java or C#.